+++ tests/recording.test.ts
@@ -2,7 +2,9 @@ import { sleep } from '@antfu/utils'
 import { AudioRecorder } from '../src'
 
 describe('class: AudioRecorder', () => {
-  const recorder = new AudioRecorder()
+  const recorder = new AudioRecorder({
+    beep: true
+  })
 
   it('can record', async () => {
     await recorder.init()
@@ -31,7 +33,11 @@ describe('class: AudioRecorder', () => {
     expect(result.duration).greaterThanOrEqual(4500)
   }, 10_000)
 
-  it('disposes properly', () => {
-    expect(recorder.stream?.active).toBe(false)
+  it('disposes properly', async () => {
+    expect(recorder.context.state).toBe<AudioContext['state']>('suspended')
+
+    await recorder.dispose()
+    expect(recorder.context.state).toBe<AudioContext['state']>('closed')
+    expect(recorder.state.inactive).toBe(true)
   })
 })